Add blocked users and user token plus changes to support these

+++ app/Http/Controllers/Api/Renders/RegistrationsController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers\Api\Renders;
 
+use App\BlockedUser;
 use App\Http\Controllers\Controller;
 use App\Render;
 use App\RenderDetail;
+use App\Team;
+use App\TeamMember;
 use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -20,7 +23,7 @@ class RegistrationsController extends Controller
     // Parameters
     const ACTIONINSTRUCTION = "actionInstruction";
     const ALLOCATEDTOUSERS = "allocatedToUsers";
-    const APITOKEN = "apiToken";
+    const USERTOKEN = "userToken";
     const AVAILABILITY = "availability";
     const C4DPROJECTWITHASSETS = "c4dProjectWithAssets";
     const C4DPROJECTNAME = "c4dProjectName";
@@ -35,7 +38,8 @@ class RegistrationsController extends Controller
     const RENDERID = "renderId";
     const SUBMISSIONSANDRENDERS = "submissionsAndRenders";
     const SUBMITTEDBYUSERID = "submittedByUserId";
-    const SUBMITTEDBYUSERAPITOKEN = "submittedByUserApiToken";
+    const SUBMITTEDBYUSERTOKEN = "submittedByUserToken";
+    const TEAMTOKEN = "teamToken";
     const TO = "to";
 
     # Action instruction
@@ -58,22 +62,31 @@ class RegistrationsController extends Controller
 
             $user = User::where('email', $request->get(self::EMAIL)) -> first();
             if ($user) {
-                $user->checkApiToken($request->get(self::APITOKEN));
+                $user->checkUserToken($request->get(self::USERTOKEN));
+
+                // Check the team token, is valid and the user is an active member
+                $teamToken = $request->get(self::TEAMTOKEN);
+                $team = Team::where('token', $teamToken) -> first();
+                if (!$team) {
+                    throw new \Exception("Team not found for token '{$teamToken}'");
+                }
+                TeamMember::checkTeamMembership($user->id, $team->id);
 
                 // Ensure there is a directory here to contain uploaded projects and renders for this user
-                $directory = "uploads/{$request->get(self::APITOKEN)}/projects";
+                $directory = "uploads/{$request->get(self::USERTOKEN)}/projects";
                 if (!file_exists($directory)) {
+                    Log::info('**** In register making directory: ' . $directory);
                     // create the directory
                     mkdir($directory, 0755, true);
                 }
-                $directory = "uploads/{$request->get(self::APITOKEN)}/renders";
+                $directory = "uploads/{$request->get(self::USERTOKEN)}/renders";
                 if (!file_exists($directory)) {
                     // create the directory
                     mkdir($directory, 0755, true);
                 }
 
                 // Update user with availability
-                $user->status     = ($request->get(self::AVAILABILITY) == self::AVAILABILITY_AVAILABLE ? 'available': 'unavailable');
+                $user->status = ($request->get(self::AVAILABILITY) == self::AVAILABILITY_AVAILABLE ? 'available': 'unavailable');
                 // Treat this as a data changed event so that the user gets the latest data
                 $user->data_changed = true;
                 $user->save();
@@ -105,6 +118,8 @@ class RegistrationsController extends Controller
     /**
      * Available notification from a slave user.
      *
+     * Test:  http://srsapi.test/api1/available?email=contact_bee@yahoo.com&teamToken=jfhdyetryetA&userToken=fl9ltqesXqPi4EkS
+     *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
@@ -120,21 +135,29 @@ class RegistrationsController extends Controller
         $to = 0;
         $outputFormat = '';
         $submittedByUserId = '';
-        $submittedByUserApiToken = '';
+        $submittedByUserToken = '';
         $renderId = '';
         try {
 
-            //Log::info('In available user for email: ' . $request->get(self::EMAIL));
+            Log::info('In available user for email: ' . $request->get(self::EMAIL));
 
             $user = User::where('email', $request->get(self::EMAIL)) -> first();
             if ($user) {
-                $user->checkApiToken($request->get(self::APITOKEN));
+                $user->checkUserToken($request->get(self::USERTOKEN));
 
-                $result = DB::table('render_details as rd')
+                // Check the team token, is valid and the user is an active member
+                $teamToken = $request->get(self::TEAMTOKEN);
+                $team = Team::where('token', $teamToken) -> first();
+                if (!$team) {
+                    throw new \Exception("Team not found for token '{$teamToken}'");
+                }
+                TeamMember::checkTeamMembership($user->id, $team->id);
+
+                $results = DB::table('render_details as rd')
                     ->select(
                         'rd.id', 'rd.status','rd.from','rd.to',
                         'r.id as render_id','r.status as render_status','r.submitted_by_user_id','r.c4dProjectWithAssets','r.c4dProjectName','r.outputFormat',
-                        'u.id as submittedByUserId','u.api_token as submittedByUserApiToken'
+                        'u.id as submittedByUserId','u.user_token as submittedByUserToken'
                     )
                     ->join('renders as r', 'r.id', '=', 'rd.render_id')
                     ->join('users as u', 'u.id', '=', 'r.submitted_by_user_id')
@@ -142,46 +165,60 @@ class RegistrationsController extends Controller
                     ->where('rd.status', RenderDetail::READY)
                     ->orderBy('render_id', 'ASC')
                     ->orderBy('rd.id', 'ASC')
-                    ->first();
+                    ->get();
 
-                if (null == $result) {
+                if (null == $results) {
                     $message = "No renders are currently pending";
                     //Log::info($message);
                     $result = 'OK';
                 } else {
-                    // Set the render to rendering
-                    $render = Render::find($result->render_id);
-                    if (!$render) {
-                        throw new \Exception("Could not find render record with id '{$result->render_id}'");
+                    foreach ($results as $result) {
+                        // If the available user is blocked, then check subsequent render details
+                        if (BlockedUser::where(['userId'=>$result->submitted_by_user_id,'teamId'=>$team->id, 'blockedUserId'=>$user->id])->exists()) {
+                            // Available user is blocked by the user who submitted the render request
+                            continue;
+                        }
+
+                        // OK, set the render to rendering
+                        $render = Render::find($result->render_id);
+                        if (!$render) {
+                            throw new \Exception("Could not find render record with id '{$result->render_id}'");
+                        }
+                        $render->status = Render::RENDERING;
+                        $render->save();
+
+                        $renderId = $result->render_id;
+                        // Allocate the render to the available slave
+                        $renderDetailId = $result->id;
+                        $renderDetail = RenderDetail::find($renderDetailId);
+                        if (!$renderDetail) {
+                            throw new \Exception("Could not find render detail record with id '{$result->id}'");
+                        }
+                        $renderDetail->status = RenderDetail::ALLOCATED;
+                        $renderDetail->allocated_to_user_id = $user->id;
+                        $renderDetail->save();
+
+                        // User's data has changed for this render
+                        Render::dataHasChanged($render->id, $user->id);
+
+                        $c4dProjectWithAssets = $result->c4dProjectWithAssets;
+                        $c4dProjectName = $result->c4dProjectName;
+                        $from = $result->from;
+                        $to = $result->to;
+                        $outputFormat = $result->outputFormat;
+                        $submittedByUserId = $result->submittedByUserId;
+                        $submittedByUserToken = $result->submittedByUserToken;
+
+                        $actionInstruction = self::AI_DO_RENDER;
+
+                        //Log::info('Instructing user for email: ' . $request->get(self::EMAIL) . ' do render detail id ' . $renderDetailId);
+
+                        break;
                     }
-                    $render->status = Render::RENDERING;
-                    $render->save();
 
-                    $renderId = $result->render_id;
-                    // Allocate the render to the available slave
-                    $renderDetailId = $result->id;
-                    $renderDetail = RenderDetail::find($renderDetailId);
-                    if (!$renderDetail) {
-                        throw new \Exception("Could not find render detail record with id '{$result->id}'");
+                    if (self::AI_DO_RENDER != $actionInstruction) {
+                        $message = "User is blocked on pending renders";
                     }
-                    $renderDetail->status = RenderDetail::ALLOCATED;
-                    $renderDetail->allocated_to_user_id = $user->id;
-                    $renderDetail->save();
-
-                    // User's data has changed for this render
-                    Render::dataHasChanged($render->id, $user->id);
-
-                    $c4dProjectWithAssets = $result->c4dProjectWithAssets;
-                    $c4dProjectName = $result->c4dProjectName;
-                    $from = $result->from;
-                    $to = $result->to;
-                    $outputFormat = $result->outputFormat;
-                    $submittedByUserId = $result->submittedByUserId;
-                    $submittedByUserApiToken = $result->submittedByUserApiToken;
-
-                    $actionInstruction = self::AI_DO_RENDER;
-
-                    //Log::info('Instructing user for email: ' . $request->get(self::EMAIL) . ' do render detail id ' . $renderDetailId);
                 }
 
                 $result = 'OK';
@@ -199,6 +236,8 @@ class RegistrationsController extends Controller
             Log::info('Error: ' . $message);
         }
 
+        Log::info('Result message: ' . $message);
+
         $returnData = [
             self::ACTIONINSTRUCTION => $actionInstruction,
             self::RENDERDETAILID => $renderDetailId,
@@ -208,7 +247,7 @@ class RegistrationsController extends Controller
             self::TO => $to,
             self::OUTPUTFORMAT => $outputFormat,
             self::SUBMITTEDBYUSERID => $submittedByUserId,
-            self::SUBMITTEDBYUSERAPITOKEN => $submittedByUserApiToken,
+            self::SUBMITTEDBYUSERTOKEN => $submittedByUserToken,
             self::RENDERID => $renderId,
             "result" => $result,
             "message" => $message,
@@ -238,7 +277,15 @@ class RegistrationsController extends Controller
 
             $user = User::where('email', $request->get(self::EMAIL)) -> first();
             if ($user) {
-                $user->checkApiToken($request->get(self::APITOKEN));
+                $user->checkUserToken($request->get(self::USERTOKEN));
+
+                // Check the team token, is valid and the user is an active member
+                $teamToken = $request->get(self::TEAMTOKEN);
+                $team = Team::where('token', $teamToken) -> first();
+                if (!$team) {
+                    throw new \Exception("Team not found for token '{$teamToken}'");
+                }
+                TeamMember::checkTeamMembership($user->id, $team->id);
 
                 // Heart beat from the slaves, it signals that the slave is active
                 //Log::info('Got user for email: ' . $request->get(self::EMAIL));
@@ -269,10 +316,15 @@ class RegistrationsController extends Controller
                                 $frameDetails[$renderId] = [];
                             }
 
+                            $dir = "uploads/{$result->submittedByUserToken}/renders/{$renderId}";
+                            if (!is_dir($dir)) {
+                                throw new \Exception("Required directory '$dir' does not exist");
+                            }
+
                             // Get all the images from the renders directory, so we can find out the actual name generated
                             $images = array_diff(
                                 scandir(
-                                    "uploads/{$result->submittedByUserApiToken}/renders/{$renderId}",
+                                    $dir,
                                     SCANDIR_SORT_DESCENDING),
                                     array('.', '..','.DS_Store')
                             );
@@ -379,7 +431,15 @@ class RegistrationsController extends Controller
 
             $user = User::where('email', $request->get(self::EMAIL)) -> first();
             if ($user) {
-                $user->checkApiToken($request->get(self::APITOKEN));
+                $user->checkUserToken($request->get(self::USERTOKEN));
+
+                // Check the team token, is valid and the user is an active member
+                $teamToken = $request->get(self::TEAMTOKEN);
+                $team = Team::where('token', $teamToken) -> first();
+                if (!$team) {
+                    throw new \Exception("Team not found for token '{$teamToken}'");
+                }
+                TeamMember::checkTeamMembership($user->id, $team->id);
 
                 $renderResults = DB::table('render_details as rd')
                     ->select(
@@ -488,7 +548,15 @@ class RegistrationsController extends Controller
             Log::info('In complete for email: ' . $request->get(self::EMAIL) . ' and render details id: ' . $request->get(self::RENDERDETAILID));
             $user = User::where('email', $request->get(self::EMAIL))->first();
             if ($user) {
-                $user->checkApiToken($request->get(self::APITOKEN));
+                $user->checkUserToken($request->get(self::USERTOKEN));
+
+                // Check the team token, is valid and the user is an active member
+                $teamToken = $request->get(self::TEAMTOKEN);
+                $team = Team::where('token', $teamToken) -> first();
+                if (!$team) {
+                    throw new \Exception("Team not found for token '{$teamToken}'");
+                }
+                TeamMember::checkTeamMembership($user->id, $team->id);
 
                 // Update the detail record to DONE
                 $renderDetailId = $request->get(self::RENDERDETAILID);
@@ -550,9 +618,17 @@ class RegistrationsController extends Controller
 
             $user = User::where('email', $request->get(self::EMAIL))->first();
             if ($user) {
-                $user->checkApiToken($request->get(self::APITOKEN));
+                $user->checkUserToken($request->get(self::USERTOKEN));
 
-                $fileNameFullPath = "uploads/{$user->api_token}/renders/{$request->get(self::RENDERID)}/{$request->get(self::FILENAME)}";
+                // Check the team token, is valid and the user is an active member
+                $teamToken = $request->get(self::TEAMTOKEN);
+                $team = Team::where('token', $teamToken) -> first();
+                if (!$team) {
+                    throw new \Exception("Team not found for token '{$teamToken}'");
+                }
+                TeamMember::checkTeamMembership($user->id, $team->id);
+
+                $fileNameFullPath = "uploads/{$user->user_token}/renders/{$request->get(self::RENDERID)}/{$request->get(self::FILENAME)}";
                 if (unlink($fileNameFullPath)) {
                     //Log::info("*** File: {$fileNameFullPath} successfully deleted");
                 } else {
@@ -587,7 +663,7 @@ class RegistrationsController extends Controller
                 ->select(
                     'r.id as render_id','r.status as render_status','r.c4dProjectWithAssets','r.outputFormat',
                     'rd.id as render_detail_id','rd.status as detail_status','rd.allocated_to_user_id','rd.from','rd.to',
-                    'u.id as submittedByUserId','u.api_token as submittedByUserApiToken'
+                    'u.id as submittedByUserId','u.user_token as submittedByUserToken'
                 )
                 ->join('render_details as rd', 'rd.render_id', '=', 'r.id')
                 ->join('users as u', 'u.id', '=', 'r.submitted_by_user_id')
@@ -609,13 +685,21 @@ class RegistrationsController extends Controller
                     //Log::info('*** Housekeeping render detail: ' . $result->render_detail_id);
 
                     // Get all the images from the renders directory, so we can delete those for this render
-                    $images = array_diff(scandir("uploads/{$result->submittedByUserApiToken}/renders", SCANDIR_SORT_DESCENDING), array('.', '..','.DS_Store'));
+                    $dir = "uploads/{$result->submittedByUserToken}/renders";
+                    if (!is_dir($dir)) {
+                        throw new \Exception("Required directory '$dir' does not exist");
+                    }
+
+                    $images = array_diff(
+                        scandir($dir, SCANDIR_SORT_DESCENDING),
+                            array('.', '..','.DS_Store')
+                    );
                     // Add an entry for each render that has occured in the range from and to
                     for ($i=$renderDetail->from; $i<=$renderDetail->to; $i++) {
                         // NB we have to find the actual name in the directory using the frame number, which is reliable.
                         $fileName = $this->getArrayValue($images, sprintf("%04d", $i) . "." . $result->outputFormat);
                         if (null !== $fileName) {
-                            if (unlink("uploads/{$result->submittedByUserApiToken}/renders/{$result->render_id}/{$fileName}")) {
+                            if (unlink("uploads/{$result->submittedByUserToken}/renders/{$result->render_id}/{$fileName}")) {
                                 Log::info("*** File: {$fileName} successfully deleted");
                             } else {
                                 throw new \Exception("Could not delete file '{$fileName}'");
